Sphinx: Build api docs as markdown
This allows us to place the api docs within the docs/ directory which allows us to directly use docusaurus versioning without additional hacks. Docusaurus versioning only versions the docs/ directory, so in the past we have resorted to bypassing docusaurus's versioning behavior and taking our own snapshot of the entire page (and updating the baseUrl to make it work).

There are some downsides here as well. Most notably, the formatting is not great. The sphinx generated docs contain a lot of detail and markdown doesn't quite capture all the styling that the html version uses.